new commit

+++ 0x0F-function_pointers/3-op_functions.c
@@ -1,66 +1,53 @@
 #include "3-calc.h"
+#include <stdlib.h>
+#include <stdio.h>
 /**
- * op_add -function for sum
- *
- * @a:integer
- *
- * @b:integer
- *
- * Return 0.
- */
+* op_add - 5 functions
+* @a: input one
+* @b: input two
+* Return: 0
+*/
 int op_add(int a, int b)
 {
 	return (a + b);
 }
 /**
- * op_sub-function to substract
- *
- * @a:integer
- *
- * @b:integer
- *
- * Return:0;
- */
+* op_sub - 5 functions
+* @a: input one
+* @b: input two
+* Return: 0
+*/
 int op_sub(int a, int b)
 {
 	return (a - b);
 }
 /**
- * op_mul-function to multiply
- *
- * @a:integer
- *
- * @b:integer
- *
- * Return:0;
- */
+* op_mul - 5 functions
+* @a: input one
+* @b: input two
+* Return: 0
+*/
 int op_mul(int a, int b)
 {
 	return (a * b);
 }
 /**
- * op_div-function that divide two numbers;
- *
- * @a:integer
- *
- * @b:integer
- *
- * Return:0;
- */
+* op_div - 5 functions
+* @a: input one
+* @b: input two
+* Return: 0
+*/
 int op_div(int a, int b)
 {
-	return ((int)(a / b));
+	return (a / b);
 }
 /**
- * op_mod-function to find a modulo of number
- *
- * @a:integer
- *
- * @b:integer
- *
- * Return:0;
- */
+* op_mod - 5 functions
+* @a: input one
+* @b: input two
+* Return: 0
+*/
 int op_mod(int a, int b)
 {
-	return ((int)(a % b));
+	return (a % b);
 }
